refactor(mj module): playing with rust modules
