+++ src/lib.rs
@@ -1,0 +1,3 @@
+mod mj;
+
+pub use mj::majority_judgment;
